[Gold III] Title: 큐브러버, Time: 680 ms, Memory: 34912 KB -BaekjoonHub

+++ 백준/Gold/3981. 큐브러버/큐브러버.py
@@ -22,9 +22,8 @@ for _ in range(T):
         f = True
         for i in range(4, len(arr)):
             j = i + 1
-            answer = Decimal(
-                a * Decimal(j ** 3) + b * Decimal(j ** 2) + c * Decimal(j) + d)
-            if abs(Decimal(arr[i]) - answer) >= 0.00001:
+            answer = a * (j ** 3) + b * (j ** 2) + c * j + d
+            if abs(arr[i] - answer) >= 0.00001:
                 f = False
 
         if f is False:
